+++ src/app/app.module.ts
@@ -1,20 +1,24 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-
 import { AppComponent } from './app.component';
-import { FirstComponentComponent } from './first-component/first-component.component';
-import { ProductComponentComponent } from './product-component/product-component.component';
+import { FirstComponent } from './first/components/first.component';
+import { CartComponent } from './cart/components/cart.component';
+import { ProductComponent } from './product/components/product/product.component';
+import { ProductlistComponent } from './product/components/products-list/productlist.component';
+import { ProductsService } from './product/services/products.service';
 
 @NgModule({
   declarations: [
     AppComponent,
-    FirstComponentComponent,
-    ProductComponentComponent
+    FirstComponent,
+    CartComponent,
+    ProductComponent,
+    ProductlistComponent
   ],
   imports: [
     BrowserModule
   ],
-  providers: [],
+  providers: [ProductsService, CartComponent],
   bootstrap: [AppComponent]
 })
 export class AppModule { }