+++ src/app/cart/services/cart.service.ts
@@ -1,0 +1,15 @@
+import { Injectable } from '@angular/core';
+import { ProductModel } from 'src/app/product/models/product.model';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class CartService {
+
+    products = [ new ProductModel(1, 'braslet', 'hv-12', 12) ];
+
+  constructor() { }
+  getProducts() {
+    return this.products;
+  }
+}