+++ src/app/first/components/product-category-enum.ts
@@ -1,0 +1,1 @@
+export enum ProductCategory { rings = 'ring', braslet = 'braslets' }