+++ src/app/product/models/product.model.ts
@@ -1,0 +1,6 @@
+export class ProductModel {
+  constructor( public id: number,
+               public name: string,
+               public sku: string,
+               public price: number ) {}
+}