+++ changelog.md
@@ -1,4 +1,6 @@
 1. Created empty AngularJewerlyProject
 2. Changed package.json start comand to lint and serve
-3. Created FirstComponent
-4.
+3. Created FirstComponent with properties, enum also included
+4. Created ProductComponent and buy button
+5. Created ProductModel, ProductService, ProductListComponent, used ngFor
+6. Created CartComponent, CartService, uased ngif+else, ngfor